+++ xmlschema/etree.py
@@ -12,9 +12,9 @@
 This module contains ElementTree setup and helpers for xmlschema package.
 """
 from __future__ import unicode_literals
-from collections import Counter
-import importlib
+import sys
 import re
+from collections import Counter
 
 import xml.etree.ElementTree as ElementTree
 
@@ -26,74 +26,42 @@ except ImportError:
 from .compat import PY3
 from .exceptions import XMLSchemaValueError, XMLSchemaTypeError
 from .namespaces import XSLT_NAMESPACE, HFP_NAMESPACE, VC_NAMESPACE
-from xmlschema.helpers import get_namespace, get_qname, qname_to_prefixed
+from .helpers import get_namespace, get_qname, qname_to_prefixed
 from .xpath import ElementPathMixin
 
 
-# Register missing namespaces into imported ElementTree module
-etree_register_namespace = ElementTree.register_namespace
-etree_register_namespace('xslt', XSLT_NAMESPACE)
-etree_register_namespace('hfp', HFP_NAMESPACE)
-etree_register_namespace('vc', VC_NAMESPACE)
+# In Python 3 the pure python implementation is hidden by the C module import,
+# so uses importlib to re-import the pure Python module, saving and restoring
+# the original modules (the optimized module and the C module).
+if PY3 and '_elementtree' in sys.modules:
+    import importlib
 
+    # Temporary remove the loaded modules
+    _pymod = sys.modules.pop('xml.etree.ElementTree')
+    _cmod = sys.modules.pop('_elementtree')
+
+    # Load the pure Python module
+    sys.modules['_elementtree'] = None
+    PyElementTree = importlib.import_module('xml.etree.ElementTree')
+
+    # Restore original modules
+    sys.modules['_elementtree'] = _cmod
+    sys.modules['xml.etree.ElementTree'] = _pymod
+else:
+    # If Python 2 or if the C module is not available
+    PyElementTree = ElementTree
+
+# ElementTree APIs
 etree_element = ElementTree.Element
 etree_parse_error = ElementTree.ParseError
 etree_parse = ElementTree.parse
 etree_iterparse = ElementTree.iterparse
 etree_fromstring = ElementTree.fromstring
+etree_register_namespace = ElementTree.register_namespace
 
-
-# Safe ElementTree
-class LazyDefusedElementTree(object):
-    """
-    Lazy importer of defused ElementTree to postpone C ElementTree override.
-    """
-    def __init__(self):
-        self._defused_etree = None
-        self.element_tree = ElementTree
-        self.etree_element = etree_element
-
-    def _import(self):
-        """Import the safe module and update set PyElementTree with the pure python module."""
-        self._defused_etree = importlib.import_module('defusedxml.ElementTree')
-        import xml.etree.ElementTree as PurePythonElementTree
-        self.element_tree = PurePythonElementTree
-        self.etree_element = PurePythonElementTree.Element
-
-    @property
-    def parse(self):
-        try:
-            return self._defused_etree.parse
-        except AttributeError:
-            self._import()
-            return self._defused_etree.parse
-
-    @property
-    def iterparse(self):
-        try:
-            return self._defused_etree.iterparse
-        except AttributeError:
-            self._import()
-            return self._defused_etree.iterparse
-
-    @property
-    def fromstring(self):
-        try:
-            return self._defused_etree.fromstring
-        except AttributeError:
-            self._import()
-            return self._defused_etree.fromstring
-
-    @property
-    def parse_error(self):
-        try:
-            return self._defused_etree.ParseError
-        except AttributeError:
-            return etree_parse_error
-
-
-defused_etree = LazyDefusedElementTree()
-
+etree_register_namespace('xslt', XSLT_NAMESPACE)
+etree_register_namespace('hfp', HFP_NAMESPACE)
+etree_register_namespace('vc', VC_NAMESPACE)
 
 # Lxml APIs
 if lxml_etree is not None:
@@ -110,6 +78,63 @@ else:
     lxml_etree_element = None
     lxml_etree_comment = None
     lxml_etree_register_namespace = None
+
+
+# Pure Python APIs
+py_etree_element = PyElementTree.Element
+py_etree_parse_error = PyElementTree.ParseError
+py_etree_register_namespace = ElementTree.register_namespace
+
+py_etree_register_namespace('xslt', XSLT_NAMESPACE)
+py_etree_register_namespace('hfp', HFP_NAMESPACE)
+py_etree_register_namespace('vc', VC_NAMESPACE)
+
+
+# Safe XML Parser and APIs
+class SafeXMLParserError(PyElementTree.ParseError):
+    pass
+
+
+class SafeXMLParser(PyElementTree.XMLParser):
+    """
+    An XMLParser that forbids entities processing.
+    """
+    def __init__(self, html=0, target=None, encoding=None):
+        super(SafeXMLParser, self).__init__(target=target, encoding=encoding)
+        parser = self.parser if PY3 else self._parser
+        parser.EntityDeclHandler = self.entity_declaration
+        parser.UnparsedEntityDeclHandler = self.unparsed_entity_declaration
+        parser.ExternalEntityRefHandler = self.external_entity_reference
+
+    def entity_declaration(self, entity_name, is_parameter_entity, value, base, system_id, public_id, notation_name):
+        raise SafeXMLParserError("Entities are forbidden (entity_name={!r})".format(entity_name))
+
+    def unparsed_entity_declaration(self, entity_name, base, system_id, public_id, notation_name):
+        raise SafeXMLParserError("Entities are forbidden (entity_name={!r})".format(entity_name))
+
+    def external_entity_reference(self, context, base, system_id, public_id):
+        raise SafeXMLParserError(
+            "External references are forbidden (system_id={!r}, public_id={!r})".format(system_id, public_id)
+        )
+
+
+def etree_safe_parse(source, parser=None):
+    if parser is None:
+        parser = SafeXMLParser(target=PyElementTree.TreeBuilder())
+    return PyElementTree.parse(source, parser)
+
+
+def etree_safe_iterparse(source, events=None, parser=None):
+    if parser is None:
+        parser = SafeXMLParser(target=PyElementTree.TreeBuilder())
+    return PyElementTree.iterparse(source, events, parser)
+
+
+def etree_safe_fromstring(text, parser=None):
+    if parser is None:
+        parser = SafeXMLParser(target=PyElementTree.TreeBuilder())
+    parser.feed(text)
+    return parser.close()
 
 
 def is_etree_element(elem):
@@ -144,11 +169,11 @@ def etree_tostring(elem, namespaces=None, indent='', max_lines=None, spaces_for_
                 etree_register_namespace(prefix, uri)
         tostring = ElementTree.tostring
 
-    elif isinstance(elem, defused_etree.etree_element):
+    elif isinstance(elem, py_etree_element):
         if namespaces:
             for prefix, uri in namespaces.items():
-                defused_etree.element_tree.register_namespace(prefix, uri)
-        tostring = defused_etree.element_tree.tostring
+                PyElementTree.register_namespace(prefix, uri)
+        tostring = PyElementTree.tostring
 
     elif lxml_etree is not None:
         if namespaces: